windows: Add windows install package
Adds an install process to replace Emoji on Windows.
It merges the EmojiOne font with Segoe UI Symbol on the user's machine,
and replaces Segoe UI Emoji as well on Win 8 and 10.

An uninstall script is also included to assist the user in restoring
the original Windows fonts.

README also updated.

Fixes #15